Use older non-queued SPI API
We really don't need the functionality of the queued SPI api because we
just queue the message as a command in our own queue and process it
(possibly) later.  This also keeps compatibility with older kernels.
I'm leaving a version check in to change when some future kernel really
does remove the old interface.