Implement gpio (minus reading input pins)
All is complete except for reading input pins.  At current, it just
returns the last known value that was written to it (if it was set to
output before) of its initial value.  We need to implement polling, but
also have some module parameter to specify rather each read should check
the last known value or query the device again.