Major re-work & correct flaw in command processing, docs
Partial re-write of command processing that corrects a nasty little
flaw. Previously, delayed commands could keep other commands from
running.  This new mechanism will ensure that commands will run in the
order they are queued with delayed commands being moved to another list
if they aren't due when popped off of the queue head and moved back to
the queue head when they are ready for execution.  This keeps command
execution order sane and still allows for all of the variants to still
work (SPI commands that need delay between URB transmission and the
configure command needing to run in a sleep-able context).

Expands documentation in Theory of Operation section and updates it so
that it now unambiguously specifies the behavior of command processing.