mcp2210-spi: fix bug rejecting xfer with zero speed
According to the documentation, struct spi_transfer::speed_hz may be
zero to use the speed from the spi_device. In calculate_spi_settings(),
this is properly calculated, but an validation step in mcp2210-spi.c's
queue_msg() function improperly rejects the message for containing such
transfers.  This patch fixes that bug.